Cleanup Auth initialization page Quickstart docs

+++ libs/Anemic/Roles.php
@@ -3,72 +3,92 @@
 declare(strict_types=1);
 
 namespace Anemic {
+
+    use SQLite3;
+
     class Roles
     {
-        static function Grant(string $id_user, string $rolename): bool
+        static function Grant(SQLite3 $db, string $email, string $rolename): bool
         {
-            return Db::RunInTx(function ($db) use ($id_user, $rolename) {
-                Db::Insert($db, "roles", [
-                    "id_user" => $id_user,
-                    "name" => $rolename
-                ]);
-            }, Config::Get("auth_database"));
+            $id_user = Users::GetId($db, $email);
+            if ($id_user === -1) {
+                return false;
+            }
+
+            return Db::Insert($db, "roles", [
+                "id_user" => $id_user,
+                "name" => $rolename
+            ]);
         }
 
-        static function Revoke(string $id_user, string $rolename): bool
+        static function GrantId(SQLite3 $db, string $id_user, string $rolename): bool
         {
-            return Db::RunInTx(function ($db) use ($id_user, $rolename) {
-                Db::Delete($db, "roles", [
-                    "id_user" => $id_user,
-                    "name" => $rolename
-                ]);
-            }, Config::Get("auth_database"));
+            return Db::Insert($db, "roles", [
+                "id_user" => $id_user,
+                "name" => $rolename
+            ]);
         }
 
-        static function GrantLogged(string $rolename): bool
+        static function Revoke(SQLite3 $db, string $email, string $rolename): bool
+        {
+            $id_user = Users::GetId($db, $email);
+            if ($id_user === -1) {
+                return false;
+            }
+            return Db::Delete($db, "roles", [
+                "id_user" => $id_user,
+                "name" => $rolename
+            ]);
+        }
+
+        static function RevokeId(SQLite3 $db, string $id_user, string $rolename): bool
+        {
+            return Db::Delete($db, "roles", [
+                "id_user" => $id_user,
+                "name" => $rolename
+            ]);
+        }
+
+
+        static function GrantLogged(SQLite3 $db, string $rolename): bool
         {
             $user = Auth::GetUser();
             if (empty($user)) {
                 return false;
             }
 
-            return Roles::Grant($user["id"], $rolename);
+            return Roles::Grant($db, $user["id"], $rolename);
         }
 
-        static function RevokeLogged(string $rolename): bool
+        static function RevokeLogged(SQLite3 $db, string $rolename): bool
         {
             $user = Auth::GetUser();
             if (empty($user)) {
                 return false;
             }
 
-            return Roles::Revoke($user["id"], $rolename);
+            return Roles::Revoke($db, $user["id"], $rolename);
         }
 
-        static function HasRoleLogged(string $rolename): bool
+        // runintx
+        static function HasRoleLogged(SQLite3 $db, string $rolename): bool
         {
             $user = Auth::GetUser();
             if (empty($user)) {
                 return false;
             }
 
-            return Roles::HasRole($user["id"], $rolename);
+            return Roles::HasRole($db, $user["id"], $rolename);
         }
 
-        static function HasRole(string $id_user, string $rolename): bool
+        static function HasRole(SQLite3 $db, string $id_user, string $rolename): bool
         {
-            $hasit = false;
+            $rs = Db::Select($db, "roles", ["id_user"], [
+                "id_user" => $id_user,
+                "name" => $rolename
+            ]);
 
-            Db::RunInTx(function ($db) use ($id_user, $rolename, &$hasit) {
-                $rs = Db::Select($db, "roles", ["id_user"], [
-                    "id_user" => $id_user,
-                    "name" => $rolename
-                ]);
-
-                $hasit = (count($rs) > 0);
-            }, Config::Get("auth_database"));
-
-            return $hasit;
+            return (count($rs) > 0);
         }
     }
 }
